+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - @imtbl/blockchain-data: Package for exposing blockchain data information.
+- @imtbl/passport: gardian domain in configuration
 
 ### Fixed
 
 - @imtbl/passport: Fixed a bug on Passport Provider getAddress.
 
 ### Changed
+- @imtbl/passport: Changed `transfer` function to call the confirmation screen with guardian check
 
 ### Removed